airflow/requirements.txt
apache-airflow==2.7.3
dbt-core
dbt-duckdb
astronomer-cosmos
